Refactor tests to pull out the mountMockedProvider functions
- Create a separate file to contain the mountMockedProvider function which is re-used in a number of tests, making it generic
